backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain_community
langchain_openai
-e .